% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourDNData.R
\name{fourDNData}
\alias{fourDNData}
\title{fourDNData}
\usage{
fourDNData(id = NULL)
}
\arguments{
\item{id}{An ID corresponding to a 4DN experiment, matching one from the
\code{fourDNDataFiles$ID} column.}
}
\value{
Local path of the queried file cached with BiocFileCache.
}
\description{
Downloads .mcool files of
4DN experiments and returns the path of the cached file.
}
\examples{
fourDNData(id = '4DNFI6CTC527')
}
