% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourDNDataFiles.R
\name{fourDNDataFiles}
\alias{fourDNDataFiles}
\title{fourDNData}
\usage{
fourDNDataFiles(experimentSetAccession = NULL, type = NULL)
}
\arguments{
\item{experimentSetAccession}{}

\item{type}{any of c('FULL', 'pairs', 'hic', 'mcool', 'boundaries',
'insulation', 'compartments')}
}
\value{
Local path of the queried file cached with BiocFileCache.
}
\description{
Fetches files from the 4DN data portal and cache them using
the BiocFileCache system.
}
\examples{
head(fourDNDataFiles)
mcf <- fourDNDataFiles(experimentSetAccession = '4DNES18BMU79', type = 'mcool')
mcf
}
