% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{fourDNDataCache}
\alias{fourDNDataCache}
\title{Manage cache / download files from the 4DN data portal}
\usage{
fourDNDataCache(...)
}
\arguments{
\item{...}{Arguments passed to internal \code{.setFourDNDataCache} function}
}
\value{
BiocFileCache object
}
\description{
Managing 4DN data downloads via the integrated
\code{BiocFileCache} system.
}
\examples{
bfc <- fourDNDataCache()
bfc
BiocFileCache::bfcinfo(bfc)
}
