% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fourDNDataFiles}
\alias{fourDNDataFiles}
\title{Set of .mcool files provided by the 4DN consortium}
\format{
An object of class \code{"data.frame"}.
}
\usage{
fourDNDataFiles
}
\description{
Set of .mcool files provided by the 4DN consortium
}
\examples{
data(fourDNDataFiles)
head(fourDNDataFiles)
}
\keyword{datasets}
